Move to mongodb

+++ src/main/java/com/zzbslayer/bookstore/datamodel/dao/CartRepository.java
@@ -3,6 +3,8 @@ package com.zzbslayer.bookstore.datamodel.dao;
 import com.zzbslayer.bookstore.datamodel.domain.Cart;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface CartRepository extends MongoRepository<Cart, Integer> {
+import java.math.BigInteger;
+
+public interface CartRepository extends MongoRepository<Cart, BigInteger> {
     Cart findByUsername(String username);
 }
